feat: SSL certs should be optional
This provider was initially written for GCP which required the SSL keys
and certificates to be supplied as values. But when extending for AWS,
the SSL certificates are installed in the OS as a certificate bundle, so
it's not necessary to supply them as values to the provider. This
changes makes it optional to specity the keys/certificates.

[#182758763](https://www.pivotaltracker.com/story/show/182758763)