feat: add support for postgres 15
[#185285503]

Postgres 15 introduced very big changes in the default permissions
for the PUBLIC scheam which completely breaks CSB functionality.

Changes in this commit restore previous functionality.
We considered other alternatives and deeply analysed the security
implications of the decision. In the end, it was reasonable to go
with pre15 functionality because CSB design assumes relation
between different bindings for the same service to be cooperative
and trustworthy.