chore: run tests introduced in story #185285503 but in main
[#185285503]

The goal of this commit is to double check if the possible bugs
identified while adding support for Postgres 15 were present in
previous builds of the provider and, ideally, in the broker.

To this end, we have cherry-picked the tests implementation and
the minimum set of code changes required to run them.
Specially, this commit doesn't include any of the code added to
support Postgres 15.