fix: role's inability to administer itself in Pg15
[#185285503]

PostgreSQL 15 removed the ability of a role to administer itself
https://github.com/postgres/postgres/commit/79de9842ab03259325ee4055fb0a7ebd2e4372ff

Apparently, CSB admin user has enough permissions to perform this
operation so we can completely remove the ConnectAsUser steps

Notice how there is an equivalent instruction in sqlUserCreate.
You could ask: "Then what's the purpose of duplicating it here?"

Because of broker upgrades.
Currently, we can't determine if an existing binding was created
with the current version of the code or an previous one which
didn't include this GRANT instruction.

It would be great to be able to recreate existing users and roles
when upgrading the broker, until then we need this as a safety net