basic prototype on how to simplify upgrade tests
Note that the changes in `csbpg/provider.go` are merely to
illustrate the point that it is working, and we get different
values depending on which version of the module we invoke.

The reason to choose the Provider function as the guinea pig for
this experiment is simply because it is one of the very few public
functions in the package. As I needed some function which was
already public in version 1.2.0 to demonstrate this method.