test: interaction of schemas and auto reassignment
Existing logic focuses mostly in simplifying and polishing
interactions with the PUBLIC schema.

However, if someone wants or needs to use other schemas the amount
and variety of side-effects grows enormously, making any attempt
to strictly manage the database an absolute pain.

This commit adds a test to demonstrate such cumbersome behaviour.