fix: some binding and permission issues
This commit includes three alternative implementations which try
to fix some existing issues in different ways.

Each of them have their pros and cons. I believe Alternative2 to
be the best solution out of the three.

The reason is that Alternative2 fixes most bugs and usability
issues but maintains a great degree of backward compatibility